Add backoff and error handling to GetEvents
A bot that received responses to the GetEvents call that were not the
heartbeat error, nor the event messages, would immetiadely requery for
events. This would cause the bot to DOS Zulip whenever it received any
kind of error, and quite hilariously, if it ever received a 429
response.

This commit adds exponential backoff and basic error handling. The
library will default to ending the EventMessage loop or channel if it
gets an error it doesn't understand, or becomes unauthorized.